% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{symplurConfig}
\alias{symplurConfig}
\title{Symplur API - Config}
\usage{
symplurConfig()
}
\description{
Replace YOURclientId and YOURclientSecret with your Symplur API credentials.
}
\examples{
symplurConfig()

}
\keyword{config}
