% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symplurContentRetweets.R
\name{symplurContentRetweets}
\alias{symplurContentRetweets}
\title{Symplur API - Content/Retweets}
\usage{
symplurContentRetweets(start = "09/01/2017", end = "09/08/2017",
  databases = "#LCSMDemoData", limit = 100)
}
\arguments{
\item{start}{Start time for period analyzed, read above API docs for more info.}

\item{end}{End time for period analyzed}

\item{databases}{The database(s) analyzed. Comma separate string if using more than one database.}

\item{limit}{Maximum number of retweets to return. Must be positive integer. Default is 10. Max is 500.}
}
\description{
API endpoint documentation:
https://api.symplur.com/v1/docs#/Twitter_Analytics:_Content/get_twitter_analytics_content_retweets
}
\examples{
LCSMDemoDataContentRetweets <- symplurContentRetweets(
    "09/01/2017",
    "09/08/2017",
    databases = "#LCSMDemoData",
    50)
}
\keyword{content}
\keyword{retweets}
