% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symplurSummary.R
\name{symplurSummary}
\alias{symplurSummary}
\title{Symplur API - Summary}
\usage{
symplurSummary(start = "09/01/2017", end = "09/08/2017",
  data = list(`databases[]` = "#LCSMDemoData"))
}
\arguments{
\item{start}{Start time for period analyzed, read above API docs for more info.}

\item{end}{End time for period analyzed}

\item{data}{The database(s) analyzed. A list.}
}
\description{
API endpoint documentation: https://api.symplur.com/v1/docs#/Twitter_Analytics:_Tweets/get_twitter_analytics_tweets_summary
}
\examples{
LCSM <- symplurSummary("09/01/2017", "09/08/2017", list("databases[]" = "#LCSMDemoData"))
}
\keyword{summary}
