% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symplurSummary.R
\name{symplurSummaryTable}
\alias{symplurSummaryTable}
\title{Symplur API - Summary Table}
\usage{
symplurSummaryTable(query = data.frame(database = character(), start =
  character(), end = character()))
}
\arguments{
\item{query}{A dataframe with columns: database, start and end.}
}
\description{
Creates a dataframe from looping through queries.
}
\details{
Example query dataframe:
\tabular{lll}{
#BCSM \tab 01/01/2010 \tab 01/01/2018\cr
#LCSM \tab 01/01/2010 \tab 01/01/2018\cr
#BTSM \tab 01/01/2010 \tab 01/01/2018
}
}
\keyword{summary}
\keyword{table}
