% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symplurTweetsActivity.R
\name{symplurTweetsActivity}
\alias{symplurTweetsActivity}
\title{Symplur API - Tweets/Activity}
\usage{
symplurTweetsActivity(start = "09/01/2017", end = "09/08/2017",
  databases = "#LCSMDemoData", interval = "days")
}
\arguments{
\item{start}{Start time for period analyzed, read above API docs for more info.}

\item{end}{End time for period analyzed}

\item{databases}{The database(s) analyzed. Comma separate string if using more than one database.}

\item{interval}{Unit of time to use for periods. Supported values are "minutes", "hours", "days", "weeks", "months", and "years". You must choose an interval that results in 500 periods or less, based on your selected date range. If omitted, the system will try to guess an appropriate interval.}
}
\description{
API endpoint documentation:
https://api.symplur.com/v1/docs#/Twitter_Analytics:_Tweets/get_twitter_analytics_tweets_activity
}
\examples{
LCSMDemoDataTweetsActivity <- symplurTweetsActivity(
    "09/01/2017",
    "09/08/2017",
    databases = "#LCSMDemoData",
    "days")
}
\keyword{activity}
\keyword{tweets}
