% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symplurTweetsSummary.R
\name{symplurTweetsSummary}
\alias{symplurTweetsSummary}
\title{Symplur API - Tweets/Summary}
\usage{
symplurTweetsSummary(start = "09/01/2017", end = "09/08/2017",
  databases = "#LCSMDemoData")
}
\arguments{
\item{start}{Start time for period analyzed, read above API docs for more info.}

\item{end}{End time for period analyzed}

\item{databases}{The database(s) analyzed. Comma separate string if using more than one database.}
}
\description{
API endpoint documentation:
https://api.symplur.com/v1/docs#/Twitter_Analytics:_Tweets/get_twitter_analytics_tweets_summary
}
\examples{
LCSMDemoDataTweetsSummary <- symplurTweetsSummary(
   "09/01/2017",
   "09/08/2017",
   databases = "#LCSMDemoData")
}
\keyword{summary}
\keyword{tweets}
